テストの共通部分の切り出し
